# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/private/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Authenticate user and generate personal access token",
                "operationId": "a60e630a5ca5c73f9c0a860da7a5399b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "vincent@test.com",
                                    "password": "M0a3r1k5"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "number",
                                                    "example": "900"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9"
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "def50200af22d480ac24fea3e0bdc8747f4"
                                                },
                                                "active_time": {
                                                    "type": "string",
                                                    "example": 1708683263
                                                },
                                                "token_expires": {
                                                    "type": "string",
                                                    "example": 1708683278
                                                },
                                                "refresh_expires": {
                                                    "type": "string",
                                                    "example": 1708683293
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 100
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "vincent mark"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "vmbcarabbacan@gmail.com"
                                                        },
                                                        "email_verified_at": {
                                                            "type": "string",
                                                            "example": null
                                                        },
                                                        "status": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2024-02-22T17:29:46.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "example": "2024-02-22T17:29:46.000000Z"
                                                        },
                                                        "status_text": {
                                                            "type": "string",
                                                            "example": "Active"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/private/user/logout/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout the user",
                "operationId": "adeeae4fd1aac65f3b866a2aa876b7e9",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "maxItems": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized access"
                                        },
                                        "data": {
                                            "properties": {
                                                "token_checker": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}